[#119] fix: layout에서 재발급이 동작하지 않아 AuthInitializer 에서 재발급 되도록 변경 및 수정
