+++ src/app/layout.tsx
@@ -4,7 +4,6 @@ import "./globals.css";
 import LayoutWrapper from "@/components/LayoutWrapper";
 import ReactQueryProvider from "./ReactQueryProvider";
 import { AuthInitializer } from "./AuthInitializer";
-import { getUserByToken, reissueServerToken } from "@/lib/auth/authFetchServer";
 
 const openSans = Open_Sans({
     variable: "--font-open-sans",
@@ -22,12 +21,6 @@ export default async function RootLayout({
 }: Readonly<{
     children: React.ReactNode;
 }>) {
-    const newAccessToken = await reissueServerToken();
-
-    let user = null;
-    if (newAccessToken) {
-        user = await getUserByToken(newAccessToken);
-    }
     return (
         <html lang="ko">
             <head>
@@ -39,18 +32,6 @@ export default async function RootLayout({
                 />
             </head>
             <body className={`${openSans.variable} antialiased `}>
-                {newAccessToken && (
-                    <script
-                        dangerouslySetInnerHTML={{
-                            __html: `
-                            window.localStorage.setItem("accessToken", "${newAccessToken}");
-                            window.localStorage.setItem("user", ${JSON.stringify(
-                                JSON.stringify(user)
-                            )});
-                        `,
-                        }}
-                    />
-                )}
                 <ReactQueryProvider>
                     <AuthInitializer>
                         <LayoutWrapper>{children}</LayoutWrapper>
