+++ src/app/mypage/page.tsx
@@ -4,21 +4,18 @@ import PageHeader from "@/components/PageHeader";
 import ChangePasswordSection from "./ChangePasswordSection";
 import DeactivateAccountSection from "./DeactivateAccountSection";
 import ProfileSection from "./ProfileSection";
-import { useRouter } from "next/navigation";
 import useMe from "@/lib/user/useMe";
 
 export default function MyPage() {
     const { data: user, isLoading } = useMe();
-    const router = useRouter();
 
     if (isLoading) {
         return <div>불러오는 중...</div>;
     }
-
-    if (!user) {
-        router.replace("/login");
-        return null;
-    }
+    // if (!user) {
+    //     router.replace("/login");
+    //     return null;
+    // }
 
     return (
         <div className="min-h-screen">
